requirements.txt
absl-py==0.5.0
asn1crypto==0.24.0
astor==0.7.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
decorator==4.3.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
http-ece==1.0.5
idna==2.7
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==3.0
Mastodon.py==1.3.1
numpy==1.14.5
protobuf==3.6.1
pycparser==2.19
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
PyYAML==3.13
requests==2.19.1
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
textgenrnn==1.4
urllib3==1.23
Werkzeug==0.14.1
